aya: remove Rc<RefCell<T>> usage in BPF program links.
This is a breaking change which transfers all ownership of a link to an
eBPF program to the caller return value of the "attach" functions. This
eliminates the overhead of Rc<RefCell<T>>, as well as prevents the
links section of ProgramData from growing without bound as programs are
attached and detached.